Split roadmap out of planner
